Refactor CI tools: split models, dedupe utilities, fix error handling
Split tools/ci/models.py:
- Move GitHub Actions schema (Step, Job, Workflow) to github_actions.py
- Add shared output helpers (format_output, bool_output, write_outputs)
- models.py now only contains internal workflow config models

Extract shared bazel-diff utilities to diff_utils.py:
- INFRA_PATTERNS, get_changed_files, has_infra_changes
- get_ci_base_commit, checkout_commit, get_or_generate_hashes
- run_bazel_diff with BazelDiffError (lets stderr pass through)

Fix is_service_running to be pure query method:
- Create separate wait_for_running method for polling behavior

Fix download_bazel_diff to raise instead of return bool:
- Errors should propagate, not be silently handled

Refactor run_release_mode to dedupe write_outputs calls:
- Extract _release_output helper and compute_release_decision function
- Single write_outputs call at the end
