+++ tools/ci/BUILD.bazel
@@ -13,6 +13,14 @@ py_library(
 )
 
 py_library(
+    name = "github_actions",
+    srcs = ["github_actions.py"],
+    imports = ["../.."],
+    visibility = ["//:__subpackages__"],
+    deps = ["@pypi//pydantic"],
+)
+
+py_library(
     name = "models",
     srcs = ["models.py"],
     imports = ["../.."],
@@ -26,12 +34,22 @@ py_library(
 # === Library modules ===
 
 py_library(
+    name = "diff_utils",
+    srcs = ["diff_utils.py"],
+    imports = ["../.."],
+    visibility = ["//:__subpackages__"],
+    deps = ["@pypi//pygit2"],
+)
+
+py_library(
     name = "bazel_diff_lib",
     srcs = ["bazel_diff_lib.py"],
     imports = ["../.."],
     visibility = ["//:__subpackages__"],
     deps = [
         ":bazel_query",
+        ":diff_utils",
+        ":github_actions",
         "//:fmt_util",
         "@pypi//pydantic",
         "@pypi//pygit2",
@@ -46,6 +64,8 @@ py_library(
     visibility = ["//:__subpackages__"],
     deps = [
         ":bazel_query",
+        ":diff_utils",
+        ":github_actions",
         ":models",
         "//:fmt_util",
         "@pypi//pydantic",
@@ -60,6 +80,7 @@ py_library(
     imports = ["../.."],
     visibility = ["//:__subpackages__"],
     deps = [
+        ":github_actions",
         ":models",
         "@pypi//pyyaml",
     ],
