+++ tools/ci/models.py
@@ -1,62 +1,18 @@
-"""Shared models for CI scripts.
+"""Workflow configuration models for CI decision logic.
 
-Contains Pydantic models used by both ci_decide.py and generate_ci.py.
+These models define our internal workflows.yaml format - trigger rules,
+inputs, and secrets for each reusable workflow.
+
+For GitHub Actions workflow schema (Step, Job, Workflow), see gha.py.
 """
 
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Annotated, Any, Literal
+from typing import Annotated, Literal
 
 import yaml
 from pydantic import BaseModel, Discriminator, Field, Tag
-
-# GitHub Actions workflow models
-
-
-class Step(BaseModel):
-    """A step in a GitHub Actions job."""
-
-    name: str | None = None
-    id: str | None = None
-    uses: str | None = None
-    run: str | None = None
-    if_cond: str | None = Field(None, alias="if", serialization_alias="if")
-    with_args: dict[str, Any] | None = Field(None, alias="with", serialization_alias="with")
-
-    model_config = {"populate_by_name": True}
-
-
-class Job(BaseModel):
-    """A job in a GitHub Actions workflow."""
-
-    name: str | None = None
-    runs_on: str | None = Field(None, alias="runs-on", serialization_alias="runs-on")
-    timeout_minutes: int | None = Field(None, alias="timeout-minutes", serialization_alias="timeout-minutes")
-    needs: str | None = None
-    if_cond: str | None = Field(None, alias="if", serialization_alias="if")
-    uses: str | None = None
-    with_args: dict[str, str] | None = Field(None, alias="with", serialization_alias="with")
-    secrets: str | None = None
-    outputs: dict[str, str] | None = None
-    steps: list[Step] | None = None
-
-    model_config = {"populate_by_name": True}
-
-
-class Workflow(BaseModel):
-    """A GitHub Actions workflow file."""
-
-    name: str
-    on: dict[str, Any] = Field(serialization_alias="on")
-    concurrency: dict[str, Any]
-    permissions: dict[str, str]
-    jobs: dict[str, Job]
-
-    model_config = {"populate_by_name": True}
-
-
-# Workflow configuration models
 
 
 class AlwaysTrigger(BaseModel):
